requirements.txt
groq
python-dotenv
flask
python-docx
PyPDF2